avoid circular dep on File
